💥 Respect "main" entry in package.json
Migration Guide:

If you previously relied on the fact that Scandium would always pick "index.js" as the entry point instead of respecting the `"main"` field of `package.json`, you need to update your `"main"` field to point to the correct file. If the `"main"` field isn't present, it still defaults to `index.js`.