💥 Add support for ESM
Migration Guide:

This shouldn't be a breaking change for most users. But the logic for detecting the entrypoint might have changed slightly. If your entrypoint is not called `index.js`, and you haven't specified a `main` file (with extension) in `package.json`, Scandium might fail to load your app. Fix it by adding a `main` field in your `package.json` file.