💥 Respect .gitignore files in subdirectories
Migration Guide:

Before this release Scandium would, in the absence of a .dockerignore file, generate a .dockerignore file based on the .gitignore file in the root. From now on, .gitignore files that are present in subdirectories will also be used to ignore files in those directories. This mimics how Git works with .gitignore files.